backend/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0

# Database dependencies
sqlalchemy==2.0.36
alembic==1.14.0
aiosqlite==0.20.0  # SQLite async driver for testing
asyncpg==0.30.0  # PostgreSQL async driver
psycopg2-binary==2.9.10  # PostgreSQL sync driver (backup)

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Google API dependencies
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.155.0

# Email dependencies
aiosmtplib==3.0.2
emails==0.6
jinja2==3.1.4

# Async and utilities
httpx==0.28.1
aiofiles==24.1.0
python-dotenv==1.0.1

# Development dependencies
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
black==24.10.0
isort==5.13.2
mypy==1.13.0

# Monitoring and logging
structlog==24.4.0
prometheus-client==0.21.1